requirements.txt
Babel==2.2.0
certifi==2015.11.20.1
cffi==1.5.2
cryptography==1.0.2
defusedxml==0.4.1
eight==0.3.3
enum34==1.1.2
Flask==0.10.1
Flask-Babel==0.9
Flask-Login==0.3.2
future==0.15.2
idna==2.0
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
lxml==3.4.4
MarkupSafe==0.23
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
pytz==2015.7
signxml==0.4.6
six==1.10.0
speaklater==1.3
Werkzeug==0.11.4
wheel==0.24.0
